Re-add required dependency (#163)
In vega-embed 3.30, vega-lite is a dependency, but in vega-embed 4.0.0 it is a devDependency. This means vega-lite now needs to be added as a separate dependency of the charts app.

I didn't notice this when testing the rennovate PRs to upgrade vega/vega-embed, as I apparently upgraded these packages without first deleting the dependencies installed by the previous version.

I included this fix in the chart-scaling PR to avoid a separate single-line PR, but @senaka thought it was a mistake and reverted it.